switching f32/f64 easily

+++ sample_main_f64.jl
@@ -2,8 +2,8 @@ using FluxMNIST
 using Flux: throttle
 
 function main(args)
-    model = FluxMNIST.Model{Float32}()
-    traindata, tX, tY = FluxMNIST.loadMNIST(Float32)
+    model = FluxMNIST.Model{Float64}()
+    traindata, tX, tY = FluxMNIST.loadMNIST(Float64)
     accuracy = FluxMNIST.Accuracy(model.m)
     evalcb = throttle(() -> @show(accuracy(tX, tY)), 10)
     @time FluxMNIST.train!(model, traindata; cb=evalcb)
